small refactor in the hook
